feat: home화면 page 및 clientPage 분리
